+++ src/client.jsx
@@ -5,42 +5,13 @@ import React from 'react';
 import {render} from 'react-dom';
 
 import {BrowserRouter} from 'react-router-dom';
+import App from './app.jsx';
 
-import {CurrentUserProvider} from './presenters/current-user.jsx';
-import {UserPrefsProvider} from './presenters/includes/user-prefs.jsx';
-import {DevTogglesProvider} from './presenters/includes/dev-toggles.jsx';
-import {Notifications} from './presenters/notifications.jsx';
-
-import Router from './presenters/pages/router.jsx';
-
-//
-// Redirects
-//
+console.log('asdf')
 if (location.hash.startsWith("#!/")) {
   window.location.replace(EDITOR_URL + window.location.hash);
+} else {
+  const dom = document.createElement('div');
+  document.body.appendChild(dom);
+  render(<BrowserRouter><App/></BrowserRouter>, dom);
 }
-
-//
-// Client side react
-//
-const App = () => (
-  <BrowserRouter>
-    <Notifications>
-      <UserPrefsProvider>
-        <DevTogglesProvider>
-          <CurrentUserProvider>
-            {api => <Router api={api}/>}
-          </CurrentUserProvider>
-        </DevTogglesProvider>
-      </UserPrefsProvider>
-    </Notifications>
-  </BrowserRouter>
-);
-
-
-//
-// Init the app
-//
-const dom = document.createElement('div');
-document.body.appendChild(dom);
-render(<App/>, dom);
